[BSK-1729] removed gcc compiler warnings about compare an `int` to ` long unsigned int` when iterating over a vector.
Former-commit-id: 00d3259d1823bbf96102a09b326cdbc09c9812a1
Former-commit-id: 24fd9cce368431d3d7449626c2f11b0d14516d7a

+++ src/simulation/dynamics/Thrusters/thrusterDynamicEffector.cpp
@@ -49,7 +49,7 @@ ThrusterDynamicEffector::ThrusterDynamicEffector()
 /*! The destructor. */
 ThrusterDynamicEffector::~ThrusterDynamicEffector()
 {
-    for (int c=0; c<this->thrusterOutMsgs.size(); c++) {
+    for (long unsigned int c=0; c<this->thrusterOutMsgs.size(); c++) {
         free(this->thrusterOutMsgs.at(c));
     }
     return;
