+++ src/simulation/environment/eclipse/eclipse.cpp
@@ -30,7 +30,7 @@ Eclipse::Eclipse()
 
 Eclipse::~Eclipse()
 {
-    for (int c=0; c<this->eclipseOutMsgs.size(); c++) {
+    for (long unsigned int c=0; c<this->eclipseOutMsgs.size(); c++) {
         delete this->eclipseOutMsgs.at(c);
     }
     return;
@@ -80,7 +80,7 @@ void Eclipse::readInputMessages()
  */
 void Eclipse::writeOutputMessages(uint64_t CurrentClock)
 {
-    for (int c = 0; c < this->eclipseOutMsgs.size(); c++) {
+    for (long unsigned int c = 0; c < this->eclipseOutMsgs.size(); c++) {
         EclipseMsgPayload tmpEclipseMsg = {};
         tmpEclipseMsg.shadowFactor = this->eclipseShadowFactors.at(c);
         this->eclipseOutMsgs.at(c)->write(&tmpEclipseMsg, this->moduleID, CurrentClock);
